Feature/tmate controller

+++ setup.py
@@ -11,15 +11,21 @@ parser = argparse.ArgumentParser()
 parser.add_argument("-c", "--clean", help="remove existing database if it exists", action="store_true")
 args = parser.parse_args()
 
-if os.path.isfile(db_path[9:]):
-    os.unlink(db_path[9:])
+if args.clean:
+    if os.path.isfile(db_path[9:]):
+        os.unlink(db_path[9:])
+    else:
+        db.drop_all()
 
-db.create_all()
+try:
+    db.create_all()
 
-session = sessionmaker(bind=create_engine(db_path))
-s = session()
-s.add(Role(name="Human"))
-s.add(Role(name="Device"))
-s.add(Role(name="Admin"))
+    session = sessionmaker(bind=create_engine(db_path))
+    s = session()
+    s.add(Role(name="Human"))
+    s.add(Role(name="Device"))
+    s.add(Role(name="Admin"))
 
-s.commit()
+    s.commit()
+except Exception as e:
+    print("Caught the following exception. Does the DB already exist?\n{}".format(e))
